pyproject.toml
[project]
name = "stockpeers"
version = "0.2.0"
description = "Stock peer analysis dashboard: easily compare stocks against others in their peer group"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "altair>=5.5.0",
    "pandas>=2.2.3",
    "streamlit>=1.44.2",
    "yfinance>=0.2.55",
    "python-decouple>=3.8",
    "python-dotenv>=1.0.0"
]
